CHE-11097: Refactor WS.NEXT applying code
Simplify KubernetesPluginsToolingApplier by moving sidecar service
discovery related code to a separate component -
SidecarServicesProvisioner.
